+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Tx/Wits.hs
@@ -16,12 +16,14 @@ module Cardano.Ledger.Api.Tx.Wits (
   bootAddrTxWitsL,
   BootstrapWitness,
 
-  -- ** Script address witness
+  -- ** Script witness
   scriptTxWitsL,
+  hashScriptTxWitsL,
 
   -- * Alonzo onwards
   AlonzoEraTxWits,
   datsTxWitsL,
+  hashDataTxWitsL,
   TxDats (..),
   unTxDats,
   rdmrsTxWitsL,
@@ -39,12 +41,13 @@ import Cardano.Ledger.Alonzo.TxWits (
   Redeemers (..),
   TxDats (..),
   datsTxWitsL,
+  hashDataTxWitsL,
   rdmrsTxWitsL,
   unRedeemers,
   unTxDats,
  )
 import Cardano.Ledger.Api.Era ()
-import Cardano.Ledger.Core (EraTxWits (..))
+import Cardano.Ledger.Core (EraTxWits (..), hashScriptTxWitsL)
 import Cardano.Ledger.Keys (KeyRole (Witness))
 import Cardano.Ledger.Keys.Bootstrap (BootstrapWitness)
 import Cardano.Ledger.Keys.WitVKey (WitVKey (WitVKey), witVKeyBytes, witVKeyHash)