+++ libs/cardano-ledger-api/CHANGELOG.md
@@ -33,6 +33,10 @@
 * Move `Cardano.Ledger.Alonzo.Tools` module from `cardano-ledegr-alonzo` into
   `Cardano.Ledger.Api.Scripts`
 * Addition of `evalBalanceTxBody`
+* Addition of `evalTxExUnits` and `evalTxExUnitsWithLogs`. Deprecated
+  `evaluateTransactionExecutionUnits` and `evaluateTransactionExecutionUnitsWithLogs` if
+  favor of the newly added ones.
+* Export `hashDataTxWitsL` and `hashScriptTxWitsL`
 
 ## 1.0.0.0