+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -22,7 +22,7 @@
 * Make `getConsumedValue` accept a deposit lookup function instead of a `DPState`
 * Add `lookupDepositDState` and `lookupRewardDState`. Former can be used with
   `getConsumedValue` to regain previous behavior.
-
+* Add `hashScriptTxWitsL`
 
 ## 1.0.0.0